consolidate modules: migrations to db, math_f32 to util; clean up wizard tests
