--- src-tauri/src/calc/mod.rs
@@ -1,2 +1,0 @@
-pub mod math_f32;
-pub mod wizard;
