+++ src-tauri/src/db/migrations.rs
@@ -6,7 +6,8 @@ use diesel_migrations::{embed_migrations, EmbeddedMigrations, MigrationHarness};
 // path configuration is in [diesel.toml]
 pub const MIGRATIONS: EmbeddedMigrations = embed_migrations!();
 
-pub fn run_migrations(
+/// Run diesel migrations. Should be called at app startup.
+pub fn run(
     connection: &mut SqliteConnection,
 ) -> Result<(), Box<dyn Error + Send + Sync + 'static>> {
     log::info!("Running migrations...");
