--- src-tauri/src/init/mod.rs
@@ -1,1 +1,0 @@
-pub mod db_setup;
