--- src-tauri/src/util/serde/mod.rs
@@ -1,1 +1,0 @@
-pub mod date;
